✨ Feat: 투표 모달 훅 구현
Resolves: #123
